major bug in signature utilization
so if for some reason the file was already instrumented
but the signature did not make it to the server on time
(happens for a plethora of reasons -- different can of worms)
then it keeps reexecuted the same file for no reason
rather we now create an empty signature.
Potential impact on fidelity? Not seen on testing?
ALso todo: if future sigs cause fetches update, update the signature
body template